requirements.txt
PySide6>=6.6
mss>=9.0.1
Pillow>=10.3.0
requests>=2.31.0
keyboard>=0.13.5
pynput>=1.7.6
pyautogui>=0.9.54